Split unit tests from bulk.rs into bulk_test.rs
Move 4 test functions (~170 lines) from the inline test module in
bulk.rs to a dedicated bulk_test.rs file. This reduces the source
file from 522 to 353 lines (32% reduction).

Snapshot files are moved to match the new module path.

Partial fix for #5116
